schemeshard: cleanup TTxState declaration, move out enums
