+++ ydb/core/tx/schemeshard/schemeshard_subop_state_types.cpp
@@ -1,0 +1,18 @@
+#include <util/generic/string.h>
+#include <util/string/printf.h>
+#include <util/string/cast.h>
+
+#include "schemeshard_subop_state_types.h"
+
+
+namespace NKikimr::NSchemeShard {
+
+TString TxStateName(ETxState s) {
+    try {
+        return ToString(s);  // auto generated by GENERATE_ENUM_SERIALIZATION
+    } catch (const std::exception& e) {
+        return Sprintf("Unknown tx-state %d", s);
+    }
+}
+
+}  // namespace NKikimr::NSchemeShard
