+++ ydb/core/tx/schemeshard/ya.make
@@ -275,11 +275,16 @@ SRCS(
     schemeshard_self_pinger.h
     schemeshard_shard_deleter.cpp
     schemeshard_shard_deleter.h
+    schemeshard_subop_state_types.cpp
+    schemeshard_subop_state_types.h
+    schemeshard_subop_types.cpp
+    schemeshard_subop_types.h
     schemeshard_svp_migration.cpp
     schemeshard_svp_migration.h
     schemeshard_system_names.cpp
     schemeshard_system_names.h
     schemeshard_sysviews_update.cpp
+    schemeshard_tx_infly.cpp
     schemeshard_tx_infly.h
     schemeshard_types.cpp
     schemeshard_types.h
@@ -290,6 +295,8 @@ SRCS(
     schemeshard_xxport__helpers.cpp
     user_attributes.cpp
 )
+
+GENERATE_ENUM_SERIALIZATION(schemeshard_subop_state_types.h)
 
 GENERATE_ENUM_SERIALIZATION(schemeshard_info_types.h)
